requirements.txt
flask==2.3.0
reportlab==4.0.7
matplotlib==3.7.1
arabic-reshaper==0.4.2
python-bidi==0.4.2
werkzeug==2.3.0